Add PlaintextCodable
