+++ Flipper/Packages/Core/Sources/Storage/Patform/PlainDeletedArchiveStorage.swift
@@ -15,7 +15,7 @@ class PlainDeletedArchiveStorage: DeletedArchiveStorage {
     func upsert(_ content: String, at path: Path) async throws {
         let path = makePath(for: path)
         print("upsert", path)
-        try storage.write(content: content, to: path)
+        try storage.write(content, at: path)
     }
 
     func delete(_ path: Path) async throws {
