+++ Flipper/Packages/Core/Sources/Storage/PlaintextCodable.swift
@@ -1,0 +1,4 @@
+protocol PlaintextCodable {
+    init(decoding: String) throws
+    func encode() throws -> String
+}
